Suggestions from review
