+++ tests/unit/pulse/test_pulse_bus_schedule.py
@@ -2,7 +2,21 @@
 import numpy as np
 import pytest
 
-from qililab.pulse import PulseBusSchedule, PulseEvent
+from qililab.pulse import Gaussian, Pulse, PulseBusSchedule, PulseEvent
+
+
+@pytest.fixture(name="mux_pulse_bus_schedule")
+def fixture_mux_pulse_bus_schedule() -> PulseBusSchedule:
+    """Return multiplexed PulseBusSchedule instance."""
+    pulse_event_1 = PulseEvent(
+        pulse=Pulse(amplitude=1, phase=0.0, duration=1000, frequency=7.0, pulse_shape=Gaussian(num_sigmas=5)),
+        start_time=0,
+    )
+    pulse_event_2 = PulseEvent(
+        pulse=Pulse(amplitude=1, phase=0.0, duration=1000, frequency=7.1, pulse_shape=Gaussian(num_sigmas=5)),
+        start_time=0,
+    )
+    return PulseBusSchedule(timeline=[pulse_event_1, pulse_event_2], port=0)
 
 
 class TestPulseBusSchedule:
@@ -31,17 +45,17 @@ class TestPulseBusSchedule:
         """Test the end_time property."""
         last_pulse_event = pulse_bus_schedule.timeline[-1]
         end = last_pulse_event.start_time + last_pulse_event.pulse.duration
-        assert end == pulse_bus_schedule.end
+        assert end == pulse_bus_schedule.end_time
 
     def test_start(self, pulse_bus_schedule: PulseBusSchedule):
         """Test the start_time property."""
         first_pulse_event = pulse_bus_schedule.timeline[0]
         start = first_pulse_event.start_time
-        assert start == pulse_bus_schedule.start
+        assert start == pulse_bus_schedule.start_time
 
     def test_total_duration(self, pulse_bus_schedule: PulseBusSchedule):
         """Test the total duration property."""
-        duration = pulse_bus_schedule.end - pulse_bus_schedule.start
+        duration = pulse_bus_schedule.end_time - pulse_bus_schedule.start_time
         assert pulse_bus_schedule.duration == duration
 
     def test_frequencies(self, mux_pulse_bus_schedule: PulseBusSchedule):
